Added a tolerance around tile for elevation interpolation.
Elevation is going to be interpolated at tiles entry and exit points,
which are computed from geodetic conversions. The result points may be
very slightly out of tiles, at numerical accuracy level. These points
should nevertheless be allowed to be interpolated using the closest
pixel data, so a tolerance was needed.

The SimpleTile class uses a fixed tolerance, arbitrarily set to 1/8
pixel.