Added a los conversion between Cartesian and geodetic coordinates.
The conversion is of course accurate only in the neighborhood of the
reference point, as a straight line in Cartesian is not a straight line
in geodetic coordinates. What is converted is the initial direction so
the two curves are tangent at the reference point.

Near Earth surface, the two curves remain within one millimeter of each
other after about 100 m travel.