Greatly improved accuracy of direct localization.
There were some small errors (at centimeter level) as result points may
be slightly out of line of sight. These errors seemed to be due to the
final linear line-of-sight model that is used at pixel level.

An iterative correction step is performed to ensure the point is really
on the line-of-sight.

The residual error is now at nanometer level.