Add and take into account atmospheric refraction flag in dump.
If the flag is set to true, we use in the DumpReplayer the default
MultiLayerModel. In case the user uses another model, one must asks the
user's model for debug purpose.
The dump files were updated to add the atmospheric refraction flag in
the direct loc and inverse loc lines.