Better management of dumped data in TilesCache when updating the tile
In the updateTile method (to be developed by the user according to the
kind of raster to be read), it happens that some useless data are
dumped. For instance, when reading some SRTM data, one needs to read
also some Geoid data. When reading the Geoid data, the elevations may be
dumped (if using the SimpleTile.interpolateElevation method).