Improved robustness of recovery algorithm when refining returns null.
Direct location is performed in two phases, with a first intersection
being computed without corrections, then correction are applied and the
first solution is "refined". It was expected that the DEM cell did not
change there. Unfortunately, sometimes the refining did not find a
result in the same DEM cell. This was a very rare occurrence, and was
detected first on very ... rugged areas like mountains.

Now, when this rare event is detected, we restart a full-blown
intersection but starting from a point just after the exit of the
failing cell, when LOS is still above DEM and we prevent the algorithm
to find solutions backward (i.e. we prevent it from finding the previous
solution again and looping infinitely).